refactor: rename JavaScript extensions to .ts
