+++ tests/json5.test.ts
